rename queryClient file for consistent case
