+++ src/components/ReviewInterface.tsx
@@ -1,7 +1,7 @@
 import { QueryClientProvider } from '@tanstack/react-query';
 import { Provider as ReduxProvider } from 'react-redux';
 import { useCurrentItem } from '#/hooks/useReactQuery';
-import { queryClient } from '#/lib/queryClient';
+import { queryClient } from '#/lib/query-client';
 import type ReviewManager from '#/lib/ReviewManager';
 import type IncrementalReadingPlugin from '#/main';
 import type ReviewView from '#/views/ReviewView';
